requirements.runtime.txt
streamlit>=1.36.0
fpdf2>=2.7.9
requests>=2.32.3
openai>=1.51.0
opentelemetry-api==1.35.0
opentelemetry-sdk==1.35.0
azure-monitor-opentelemetry-exporter>=1.0.0b35
azure-storage-blob>=12.23.1
azure-identity>=1.17.1
azure-ai-projects>=2.0.0b4
